Update motorbike job-scripts
Update the job-scripts and jobbergate's job-script templates for the
motorbike example.

Set option --bind $PWD:$HOME in singularity exec command to bind
current dir to the current user home dir in the container. It is needed
when running inside /nfs dir.

Download OPENFOAM_DIR and SINGULARITY_IMAGE only if they are
not available yet.